functions/requirements.txt
boto3
python-jose